+++ Sidekiq Z2/CTranslation/main.c
@@ -17,17 +17,17 @@ int main(){
     char scheme[] = "BPSK"; // Modulation scheme 'OOK', 'QPSK', or 'QAM'
     double alpha = 0.5; // roll-off factor of the RRC pulse-shaping filter
 
-    struct Array_Tuple preamble = defineArray((int[]){0,1,0,0,0,0,1,1,0,0,0,1,0,1,0,0,1,1,1,1,0,1,0,0,0,1,1,1,0,0,1,0,0,1,0,1,1,0,1,1,1,0,1,1,0,0,1,1,0,1,0,1,0,1,1,1,1,1,1,0}, 60);// optimal periodic binary code for N = 63 https://ntrs.nasa.gov/citations/19800017860
+    struct Array_Tuple preamble = defineArray((double[]){0,1,0,0,0,0,1,1,0,0,0,1,0,1,0,0,1,1,1,1,0,1,0,0,0,1,1,1,0,0,1,0,0,1,0,1,1,0,1,1,1,0,1,1,0,0,1,1,0,1,0,1,0,1,1,1,1,1,1,0}, 60);// optimal periodic binary code for N = 63 https://ntrs.nasa.gov/citations/19800017860
     struct Array_Tuple data = {randomArray(2,data_length), data_length}; // this points to the random array generated;  creates 256 random bits
 
-    struct Array_Tuple CRC_key = defineArray((int[]){1,0,0,1,1,0,0,0,0,1,1,1}, 12);// Best CRC polynomials: https://users.ece.cmu.edu/~koopman/crc/
+    struct Array_Tuple CRC_key = defineArray((double[]){1,0,0,1,1,0,0,0,0,1,1,1}, 12);// Best CRC polynomials: https://users.ece.cmu.edu/~koopman/crc/
     
     
     // Working Demo Code Here ----------------------------------------------------
     // calloc(length, size) is used in defineArray() and for the output variable of CRC_encodeData()
 
-    struct Array_Tuple demoA = defineArray((int[]){1,0,0,1,0,0}, 6);
-    struct Array_Tuple demoB = defineArray((int[]){1,1,0,1}, 4);
+    struct Array_Tuple demoA = defineArray((double[]){1,0,0,1,0,0}, 6);
+    struct Array_Tuple demoB = defineArray((double[]){1,1,0,1}, 4);
     struct Array_Tuple out = CRC_encodeData(demoA, demoB);
     
     printArray("Out", out.array, out.length);
